rename module storage to manager
